refactor: move datetime helpers into controller file
